readelf: improve sh_link validation
Some sh_link uses did not include an out of range check. Add those, and
move the warning to load_sections.

We could instead just exit() in load_sections for an out of range
sh_link, but for this tool I think it is better to report as much
information as we can parse.

Found with afl 1.83b.